requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
redis==5.0.1
docker==7.0.0
paho-mqtt==1.6.1
httpx==0.26.0
pydantic==2.6.1
python-dotenv==1.0.1
aiosqlite==0.19.0
psutil==5.9.8
